+++ src/main/java/gift/oauth/service/DefaultKakaoAuthService.java
@@ -1,15 +1,18 @@
-package gift.auth.service;
+package gift.oauth.service;
 
-import gift.auth.client.KakaoOAuthClient;
-import gift.auth.dto.KakaoTokenResponseDto;
-import gift.auth.dto.KakaoUserInfoResponseDto;
+import gift.oauth.client.KakaoOAuthClient;
+import gift.oauth.dto.KakaoLoginResponseDto;
+import gift.oauth.dto.KakaoTokenResponseDto;
+import gift.oauth.dto.KakaoUserInfoResponseDto;
 import gift.common.security.JwtUtil;
-import gift.member.dto.MemberResponseDto;
 import gift.member.entity.Member;
 import gift.member.entity.Role;
 import gift.member.repository.MemberRepository;
+import gift.oauth.entity.UserKakaoToken;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.time.Instant;
 
 @Service
 public class DefaultKakaoAuthService implements KakaoAuthService {
@@ -30,7 +33,7 @@ public class DefaultKakaoAuthService implements KakaoAuthService {
 
     @Override
     @Transactional
-    public MemberResponseDto loginWithKakao(String code) {
+    public KakaoLoginResponseDto loginWithKakao(String code) {
 
         KakaoTokenResponseDto tokenResponse = kakaoOAuthClient.requestToken(code);
 
@@ -42,9 +45,22 @@ public class DefaultKakaoAuthService implements KakaoAuthService {
                     return memberRepository.save(newMember);
                 });
 
+        Instant now = Instant.now();
+        Instant accessTokenExpiresAt = now.plusSeconds(tokenResponse.expiresIn());
+        Instant refreshTokenExpiresAt = now.plusSeconds(tokenResponse.refreshTokenExpiresIn());
+
+        UserKakaoToken kakaoToken = member.getKakaoToken();
+        if (kakaoToken == null) {
+            kakaoToken = new UserKakaoToken(member, tokenResponse.accessToken(), tokenResponse.refreshToken(),
+                    accessTokenExpiresAt, refreshTokenExpiresAt);
+            member.setKakaoToken(kakaoToken);
+        } else {
+            kakaoToken.update(tokenResponse.accessToken(), accessTokenExpiresAt);
+        }
+
         // JWT 발급
         String jwt = jwtUtil.generateToken(member.getEmail(), member.getId(), member.getRole().name());
 
-        return new MemberResponseDto(member);
+        return new KakaoLoginResponseDto(jwt, tokenResponse);
     }
 }
